Give Symbol pub fields and remove getters and setters
This help with identifying where clones actually need to happen and allows for taking references to the fields. It also removes the need for various getters such as `get_args`, `get_args_mut` and `get_args_owned`. The downside is that more clones appear in the code. However, these can be removed if the rest of the code uses more references too.