Allow running tests with custom output writer
This implements a new `run_tests_with_writer()` method for `Module` (and
a wrapper around it for `Compiled`), which takes an additional argument:
an `std::io::Write` impl. This can be used to write the test output to
something other than standard output.

The `run_tests()` function is kept compatible, as a wrapper around the
new function, which sets the writer to `std::io::stdout()`.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>