Allow writing custom test runners
This introduces a new `.get_tests()` method for `Module` (and a wrapper
around it for `Compiled`), which returns an iterator over
`codegen::TestCase`, a helper struct that encapsulates the test's name,
and the compiled function.

This iterator can be used to implement custom runners that filter tests,
run them with independent contexts, or provide a different kind of
output than the built in `run_tests()` method does, and so on.

The existing `run_tests()` method was also changed to use
`.get_tests()`.

Test cases to exercise the new method are included.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>